[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8673 Former-commit-id: 2cfab1ef690d1b39ae97e54fbefbe1de3f90702f

+++ icu4c/source/common/ucnv_io.c
@@ -577,7 +577,7 @@ ucnv_io_setDefaultConverterName(const char *converterName) {
             defaultConverterName=name;
         } else {
             /* do not set the name if the alias lookup failed and it is too long */
-            int32_t length=uprv_strlen(converterName);
+            int32_t length=(int32_t)(uprv_strlen(converterName));
             if(length<sizeof(defaultConverterNameBuffer)) {
                 /* it was not found as an alias, so copy it - accept an empty name */
                 UBool didLock;
